Виправлено роботу 6 на Python

+++ lab_op_6_python.py
@@ -3,8 +3,8 @@ def distance(x,y):
 
 def cycle(dist_near):
     for i in range(2, count + 1):
-        x = int(input('x{}: '.format(i)))
-        y = int(input('y{}: '.format(i)))
+        x = float(input('x{}: '.format(i)))
+        y = float(input('y{}: '.format(i)))
         dist_curr = distance(x, y)
         if dist_curr < dist_near:
             dist_near = dist_curr
@@ -13,8 +13,8 @@ def cycle(dist_near):
 
 i = 1
 count = int(input('how many points?:'))
-x = int(input('x{}: '.format(i)))
-y = int(input('y{}: '.format(i)))
+x = float(input('x{}: '.format(i)))
+y = float(input('y{}: '.format(i)))
 dist_near = distance(x, y)
 ind_near = 1
 cycle(dist_near)
